pyproject.toml
[tool.poetry]
name = "files-merge"
version = "0.1.0"
description = ""
authors = ["Mike <1651319@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
databases = {extras = ["asyncpg"], version = "^0.9.0"}
python-dotenv = "^1.0.1"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
